Import 5.x MELSpace theme
This lives in the root of the repository because it is copied to the
the xmlui-mirage2 overlay during Docker build depending on an envir-
onment variable controlling the currently active theme. I copy it to
xmlui-mirage2 now so I can build and test it without Docker.

Latest upstream commit: c13a89176fb5d15e3781abea0cf22c3f462b8dbe