Dockerfile: Adjust mvn invocation
We need to tell the build system that we already have the build env
set up correctly so that it doesn't download local Node.js and Ruby
copies during the build process. The -Dmirage2.deps.included=false
option is not very obvious because "included" here means that they
will be INCLUDED during build time.