rootfs/dspace/bin/setup-postgres: Improve substitution
In my testing I was using 'dspacetest' as a database name and pass,
which caused this search and replace to become "dspacetesttest". We
should be anchoring the search pattern with ^ and $, but that means
we need to change the sed delimiter from "#" to something else like
"," so that the resulting "$#" doesn't have special meaning to the
shell (in this case $# is the number of arguments).